fix: resolve clippy warnings and axum version conflict
- Upgrade axum-test to 15.7.4 for axum 0.8 compatibility
- Fix test headers to use HeaderName/HeaderValue types
- Fix sort_by to sort_by_key with Reverse for memory descending
- Fix CPU load casts (load.one, load.five, load.fifteen are already f64)
- Fix Docker: remove unnecessary cast, fix unwrap_or_else, use networks.values()
- Remove unused imports and dead code in tests
- Fix sort_by_key for PID ascending and Name sorting
- cargo fmt passes

+++ agent/src/metrics/cpu.rs
@@ -37,9 +37,9 @@ impl CpuCollector {
 
         let load = sysinfo::System::load_average();
         let load_average = LoadAverage {
-            one: load.one as f64,
-            five: load.five as f64,
-            fifteen: load.fifteen as f64,
+            one: load.one,
+            five: load.five,
+            fifteen: load.fifteen,
         };
 
         Ok(CpuMetrics {
